local models: fileSha256 and verifyModelFile move to modelVerify.ts
The MLX downloader needs them, and localModels.ts will import the downloader.
