+++ packages/agency-lang/lib/stdlib/localModels.ts
@@ -4,15 +4,12 @@ import {
   stat,
   list,
   remove,
-  move,
   readText,
-  readStream,
   writeText,
 } from "./contained.js";
 import * as os from "node:os";
 import * as path from "node:path";
 import { fileURLToPath } from "node:url";
-import { createHash } from "node:crypto";
 import { z } from "zod";
 import { findFileUp } from "../importPaths.js";
 import {
@@ -22,6 +19,8 @@ import {
 } from "../runtime/localProvider.js";
 import { __ctx } from "../runtime/asyncContext.js";
 import { recordDownload, readDownloadManifest } from "./localModelManifest.js";
+import { fileSha256, verifyModelFile } from "./modelVerify.js";
+export { fileSha256, verifyModelFile } from "./modelVerify.js";
 import {
   type Backend,
   isGgufPath,
@@ -1197,63 +1196,6 @@ export async function _registerLocalProvider(): Promise<void> {
     entryPath: choice.entryPath,
     entrySource: choice.source,
   });
-}
-
-/** Stream-hash a file's SHA-256 (hex), never buffering the whole file. The
- *  `update` is guarded so a synchronous throw in the data handler rejects the
- *  promise instead of escaping it. */
-export function fileSha256(filePath: string): Promise<string> {
-  return new Promise((resolve, reject) => {
-    const hash = createHash("sha256");
-    let stream: ReturnType<typeof readStream>;
-    try {
-      const located = wholePath(filePath);
-      stream = readStream(located.root, located.target);
-    } catch (err) {
-      reject(err as Error);
-      return;
-    }
-    stream.on("error", reject);
-    stream.on("data", (chunk) => {
-      try {
-        hash.update(chunk);
-      } catch (err) {
-        stream.destroy();
-        reject(err as Error);
-      }
-    });
-    stream.on("end", () => resolve(hash.digest("hex")));
-  });
-}
-
-/** Verify `filePath` against the expected hex SHA-256. On mismatch, rename the
- *  file to `<filePath>.invalidSha` (kept for inspection; won't be picked up, so
- *  the next run re-downloads) and throw. A failed rename is logged (not
- *  swallowed) and reflected in the thrown message. */
-export async function verifyModelFile(
-  filePath: string,
-  expected: string,
-  name: string,
-): Promise<void> {
-  // `fileSha256` returns lowercase hex; normalize the expected pin so a
-  // valid-but-uppercase hash (e.g. from a hand-written alias) still matches.
-  const want = expected.toLowerCase();
-  const actual = await fileSha256(filePath);
-  if (actual === want) return;
-  const quarantine = `${filePath}.invalidSha`;
-  let moved = true;
-  try {
-    move(wholePath(filePath), wholePath(quarantine));
-  } catch (err) {
-    moved = false;
-    console.warn(`Could not move "${filePath}" to "${quarantine}" after SHA-256 mismatch:`, err);
-  }
-  throw new Error(
-    `SHA-256 verification failed for "${name}": expected ${expected}, got ${actual}. ` +
-      (moved
-        ? `The downloaded file was moved to ${quarantine} for inspection and will be re-downloaded next time.`
-        : `The downloaded file at ${filePath} could NOT be moved aside — delete it manually before re-running.`),
-  );
 }
 
 /** The pinned SHA-256 for a model name/alias, or undefined when none is known
